Adding colors
