+++ rft-kocsmapp/app/src/main/AndroidManifest.xml
@@ -23,10 +23,6 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".FacebookLoginActivity"
-            android:label="@string/title_activity_facebook_login"
-            android:parentActivityName=".MainActivity" />
-        <activity
             android:name=".ShowPubsAcivity"
             android:parentActivityName=".MainActivity" />
         <activity
